requirements.txt
gradio>=4.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
wikipedia>=1.4.0
python-dotenv>=1.0.0
